Move each class definition to its own file for better code organization
