+++ projects/poo-videojuego-rpg/src/models/characters.ts
@@ -93,90 +93,5 @@ export abstract class Hero {
     } else {
       console.log("Lo lamento, no has ganado ningún ataque en esta ronda :(");
     }
-
   }
-
-  abstract levelUp(): void;
-}
-
-export class Wizard extends Hero {
-  private magic: number;
-
-  constructor(name: string) {
-    super(name);
-    this.magic = 25;
-  }
-
-  // Polimorfismo y sobrecarga de metodo
-  attack(skill?: Skill): void {
-    if (skill) {
-      if (this.magic < skill.cost) {
-        console.log(
-          `${this.name} no tiene la magia necesaria para usar el ataque: ${skill.name}`
-        );
-        return;
-      }
-      this.magic -= skill.cost;
-      this.xp += skill.xp;
-      console.log(`${this.name} esta atacando con: ${skill.name}`);
-    } else {
-      console.log(`${this.name} esta haciendo un ataque básico...`);
-    }
-  }
-
-  levelUp(): void {}
-}
-
-export class Archer extends Hero {
-  private arrows: number;
-
-  constructor(name: string) {
-    super(name);
-    this.arrows = 25;
-  }
-  // Polimorfismo y sobrecarga de metodo
-  attack(skill?: Skill): void {
-    if (skill) {
-      if (this.arrows < skill.cost) {
-        console.log(
-          `${this.name} no tiene la cantidad necesaria de flechas para usar el ataque: ${skill.name}`
-        );
-        return;
-      }
-      this.arrows -= skill.cost;
-      this.xp += skill.xp;
-      console.log(`${this.name} esta atacando con: ${skill.name}`);
-    } else {
-      console.log(`${this.name} esta haciendo un ataque básico...`);
-    }
-  }
-
-  levelUp(): void {}
-}
-
-export class Fighter extends Hero {
-  private strength: number;
-
-  constructor(name: string) {
-    super(name);
-    this.strength = 25;
-  }
-  // Polimorfismo y sobrecarga de metodo
-  attack(skill?: Skill): void {
-    if (skill) {
-      if (this.strength < skill.cost) {
-        console.log(
-          `${this.name} no tiene la fuerza necesaria para usar el ataque: ${skill.name}`
-        );
-        return;
-      }
-      this.strength -= skill.cost;
-      this.xp += skill.xp;
-      console.log(`${this.name} esta atacando con: ${skill.name}`);
-    } else {
-      console.log(`${this.name} esta haciendo un ataque básico...`);
-    }
-  }
-
-  levelUp(): void {}
 }
